requirements.txt
pipenv

django
Djangorestframework
djoser
django-heroku

psycopg2-binary
gunicorn
